research doc

+++ src/main/java/com/example/easywheelz/buisness/impl/car/DeleteCarUseCaseImpl.java
@@ -1,5 +1,6 @@
 package com.example.easywheelz.buisness.impl.car;
 
+import com.example.easywheelz.Errors.InvalidCarCredentials;
 import com.example.easywheelz.buisness.interfaces.car.DeleteCarUseCase;
 import com.example.easywheelz.persistance.CarRepository;
 import lombok.AllArgsConstructor;
@@ -11,6 +12,12 @@ public class DeleteCarUseCaseImpl implements DeleteCarUseCase{
     private final CarRepository carRepository;
     @Override
     public void deleteCar(long carId) {
-        carRepository.deleteById(carId);
+        if (carRepository.existsById(carId)){
+            carRepository.deleteById(carId);
+
+        }
+        else {
+            throw new InvalidCarCredentials("Car not found");
+        }
     }
 }
